Converted CritterDetails.cs from uint to ushort

+++ Engine/Province/Critters/CritterBuilder/ClassRatio/IClassRatio.cs
@@ -2,5 +2,5 @@
 
 public interface IClassRatio
 {
-    public (uint Lower, uint Middle, uint Upper) Execute(uint count);
+    public (ushort Lower, ushort Middle, ushort Upper) Execute(ushort count);
 }
